+++ src/sklearndf/wrapper/numpy/_numpy.py
@@ -1,5 +1,5 @@
 """
-Estimator adapter classes to handle numpy arrays in meta-estimators
+Estimator adapter classes to handle numpy arrays in meta-estimators.
 """
 
 import logging
@@ -9,7 +9,7 @@ import numpy as np
 import numpy.typing as npt
 import pandas as pd
 
-from pytools.api import AllTracker, inheritdoc
+from pytools.api import AllTracker, inheritdoc, subsdoc
 
 from sklearndf import (
     ClassifierDF,
@@ -59,9 +59,8 @@ class EstimatorNPDF(
     Generic[T_DelegateEstimatorDF],
 ):
     """
-    An adapter class that wraps around a :class:`.EstimatorDF` and accepts numpy arrays
-    for all DF estimator methods that would usually only accept pandas data frames or
-    series.
+    Adapter class that delegates to :class:`.EstimatorDF` and accepts numpy arrays
+    in addition to :mod:`pandas` data frames and series.
 
     Converts all numpy arrays to pandas series or data frames before deferring to the
     delegate estimator, and passes through pandas objects unchanged.
@@ -163,10 +162,15 @@ class EstimatorNPDF(
 
 # noinspection PyPep8Naming
 @inheritdoc(match="""[see superclass]""")
+@subsdoc(pattern="EstimatorDF", replacement="LearnerDF", using=EstimatorNPDF)
 class LearnerNPDF(
     EstimatorNPDF[T_DelegateLearnerDF], LearnerDF, Generic[T_DelegateLearnerDF]
 ):
     """[see superclass]"""
+
+    # repeating attribute declarations of base classes for Sphinx documentation
+    delegate: T_DelegateLearnerDF
+    column_names: Optional[Union[Sequence[str], Callable[[], Sequence[str]]]]
 
     def predict(
         self, X: Union[npt.NDArray[Any], pd.DataFrame], **predict_params: Any
@@ -177,12 +181,17 @@ class LearnerNPDF(
 
 # noinspection PyPep8Naming
 @inheritdoc(match="""[see superclass]""")
+@subsdoc(pattern="EstimatorDF", replacement="SupervisedLearnerDF", using=EstimatorNPDF)
 class SupervisedLearnerNPDF(
     LearnerNPDF[T_DelegateSupervisedLearnerDF],
     SupervisedLearnerDF,
     Generic[T_DelegateSupervisedLearnerDF],
 ):
     """[see superclass]"""
+
+    # repeating attribute declarations of base classes for Sphinx documentation
+    delegate: T_DelegateSupervisedLearnerDF
+    column_names: Optional[Union[Sequence[str], Callable[[], Sequence[str]]]]
 
     def score(
         self,
@@ -200,12 +209,17 @@ class SupervisedLearnerNPDF(
 
 # noinspection PyPep8Naming
 @inheritdoc(match="""[see superclass]""")
+@subsdoc(pattern="EstimatorDF", replacement="ClassifierDF", using=EstimatorNPDF)
 class ClassifierNPDF(
     SupervisedLearnerNPDF[T_DelegateClassifierDF],
     ClassifierDF,
     Generic[T_DelegateClassifierDF],
 ):
     """[see superclass]"""
+
+    # repeating attribute declarations of base classes for Sphinx documentation
+    delegate: T_DelegateClassifierDF
+    column_names: Optional[Union[Sequence[str], Callable[[], Sequence[str]]]]
 
     @property
     def classes_(self) -> Union[npt.NDArray[Any], List[npt.NDArray[Any]]]:
@@ -243,6 +257,7 @@ class ClassifierNPDF(
 
 # noinspection PyPep8Naming
 @inheritdoc(match="""[see superclass]""")
+@subsdoc(pattern="EstimatorDF", replacement="RegressorDF", using=EstimatorNPDF)
 class RegressorNPDF(
     SupervisedLearnerNPDF[T_DelegateRegressorDF],
     RegressorDF,
@@ -250,15 +265,24 @@ class RegressorNPDF(
 ):
     """[see superclass]"""
 
+    # repeating attribute declarations of base classes for Sphinx documentation
+    delegate: T_DelegateRegressorDF
+    column_names: Optional[Union[Sequence[str], Callable[[], Sequence[str]]]]
+
 
 # noinspection PyPep8Naming
 @inheritdoc(match="""[see superclass]""")
+@subsdoc(pattern="EstimatorDF", replacement="TransformerDF", using=EstimatorNPDF)
 class TransformerNPDF(
     EstimatorNPDF[T_DelegateTransformerDF],
     TransformerDF,
     Generic[T_DelegateTransformerDF],
 ):
     """[see superclass]"""
+
+    # repeating attribute declarations of base classes for Sphinx documentation
+    delegate: T_DelegateTransformerDF
+    column_names: Optional[Union[Sequence[str], Callable[[], Sequence[str]]]]
 
     def transform(self, X: Union[npt.NDArray[Any], pd.DataFrame]) -> pd.DataFrame:
         """[see superclass]"""