+++ src/sklearndf/wrapper/stacking/__init__.py
@@ -1,0 +1,5 @@
+"""
+DF wrapper classes for stacking estimators.
+"""
+
+from ._stacking import *